+++ UI/obs-app.cpp
@@ -495,8 +495,6 @@ void OBSApp::InitUserConfigDefaults()
 	config_set_default_bool(userConfig, "BasicWindow", "SpacingHelpersEnabled", true);
 	config_set_default_bool(userConfig, "BasicWindow", "RecordWhenStreaming", false);
 	config_set_default_bool(userConfig, "BasicWindow", "KeepRecordingWhenStreamStops", false);
-	config_set_default_bool(userConfig, "BasicWindow", "StatusOverlayEnabled", false);
-	config_set_default_string(userConfig, "BasicWindow", "StatusOverlayPosition", "BottomRight");
 	config_set_default_bool(userConfig, "BasicWindow", "SysTrayEnabled", true);
 #ifdef OBS_AMD_LITE
 	config_set_default_bool(userConfig, "BasicWindow", "SysTrayWhenStarted", false);