+++ UI/window-basic-main.hpp
@@ -57,7 +57,6 @@ class QListWidgetItem;
 class VolControl;
 class OBSBasicStats;
 class OBSBasicVCamConfig;
-class OBSStatusOverlay;
 
 #include "ui_OBSBasic.h"
 #include "ui_ColorSelect.h"
@@ -292,7 +291,6 @@ private:
 	QPointer<QTimer> diskFullTimer;
 
 	QPointer<QTimer> nudge_timer;
-	std::unique_ptr<OBSStatusOverlay> statusOverlay;
 	bool recent_nudge = false;
 
 	os_cpu_usage_info_t *cpuUsageInfo = nullptr;
@@ -900,7 +898,6 @@ public:
 	OBSScene GetCurrentScene();
 
 	void SysTrayNotify(const QString &text, QSystemTrayIcon::MessageIcon n);
-	void UpdateStatusOverlaySettings();
 
 	inline OBSSource GetCurrentSceneSource()
 	{
@@ -1024,8 +1021,6 @@ public:
 	void CreateFilterPasteUndoRedoAction(const QString &text, obs_source_t *source, obs_data_array_t *undo_array,
 					     obs_data_array_t *redo_array);
 
-	void InitStatusOverlay();
-	void SyncStatusOverlayState(const QString &flashText = QString());
 	void SetDisplayAffinity(QWindow *window);
 
 	QColor GetSelectionColor() const;